$DOMAIN/requirements/dev.txt
-r common.txt
coverage==3.7.1
django_debug_toolbar==1.2.1
factory-boy==2.4.1
pip-tools==0.3.4
